Parse global switches before opening the HID dev
Due to how everything is currently structured, we need to open the
keyboard device before parsing the effects and their parameters, because
we just interpret them as we parse them.

However, that means that if you given anything like --help (which is
currently just interpreted by means of “Unrecognized effect”), leddy
will still try to open the HID device, which may just fail if you don’t
run it as root.

Leddy shouldn’t require elevated access rights just to print some --help
text, and maybe in the future we also want to allow users to specify the
USB vendor:product ID combination, or select one of multiple keyboards,
or something like that; and all of that should happen before the
keyboard device is opened.

So introduce a primary parsing round for global switches that do not
require device access, and that we can thus execute before attempting to
connect to the device.