main: Sort gradient colors by position
Turns out the keyboard always strictly interprets the colors in the
order they arrive.  That means we should reorder them by position to get
a reasonable result.  (But use a stable sort, because if the user gives
colors at the same position, they should still be ordered the same way
they were given in the list.)