Fixed perm change; allow per-key color
With the newest insights (see HEAD^, or rather HEAD:doc.txt), it becomes
clear that to make a change permanent for some profile, we need to send:
  0x05 <profile> 0x02 <effect>
  0x13
  0x04 <profile>

Instead of
  0x0f <effect>
  0x04 <some garbage>

And that seems to work indeed.

With this, we can now allow changing per-key colors based on a list
received from stdin.

Fixes: #2
Fixes: #4