sound_spectrum: Various improvements
For one, the colors are smoother now that every bar has a single color
that is just interpolated between the different octaves.

Second, the three colors are now three octaves.

Third, the bottom row is now solely bass.

Fourth, scaling is now dynamic.

Fifth, the arrow keys now show the combined bass compared to the rest of
the spectrum.

Sixth, the navigation keys show the dominant frequency (color), its
relationship to the rest of the spectrum (saturation), and its intensity
(value).